requirements.txt
matplotlib==3.10.0
numpy==1.23.1
pandas==2.2.3
plotly==5.24.1
psycopg2==2.9.3
psycopg2_binary==2.9.3
psycopg2_binary==2.9.10
python-dotenv==1.0.1
scikit_learn==1.3.0
scipy==1.15.1
streamlit==1.41.1
